\name{quantileVaR-package}
\alias{quantileVaR-package}
\alias{quantileVaR}
\docType{package}
\title{
Quantile regression for estimating Value at Risk and backtesting.
}

\description{
Estimating CAViaR models (Engle & Manganelli, 2002) and similar models. Create VaR forecasts. Backtest VaR forecasts (independent of model used for creating the forecasts).
}
\details{
\tabular{ll}{
Package: \tab quantileVaR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-08-14\cr
License: \tab GPL-2 | GPL-3\cr
}

}
\author{
Steinar Veka

Maintainer: Who to complain to <steinar.veka@gmail.com>
}
\references{
Robert F. Engle and Simone Manganelli
CAViaR: Conditional Autoregressive Value at Risk by Regression Quantiles
Journal of Business & Economic Statistics
Vol. 22, No. 4 (Oct., 2004), pp. 367-381
Published by: American Statistical Association
Article Stable URL: http://www.jstor.org/stable/1392044

Peter F. Christoffersen
Evaluating Interval Forecasts
International Economic Review
Vol. 39, No. 4, Symposium on Forecasting and Empirical Methods in Macroeconomics and Finance (Nov., 1998), pp. 841-862
Published by: Wiley
Article Stable URL: http://www.jstor.org/stable/2527341
}
